Move reindex rule in indexer init function
